refactor: share script create/delete CRUD across published and draft routes
The per-file create and delete handlers duplicated all their substance
between PublishedAssignmentRoutes+ScriptCRUD and
DraftAssignmentRoutes+SuiteEditing: filename sanitization, duplicate-name
and pattern-family/dependent guards, assignment-scope variable inlining,
zip write/remove, and manifest add/remove. Extract createScriptInSetup and
deleteScriptFromSetup into a new ScriptCRUDHelpers, with a shared
CreateScriptBody decode struct.

The published handlers retain their support-file re-extraction to the
shared student directory and the editURL in the response; the draft
handlers, which have neither students nor a stable assignment route yet,
retain neither — that divergence is now explicit rather than buried in two
near-identical copies.

Net ~130 fewer lines in the handlers. Build, swift-format, SwiftLint
--strict, and the script/suite route test suites (135 tests) pass.
